refactor audio helper

+++ app/AudioHelper.py
@@ -4,25 +4,29 @@ from pyaudio import PyAudio, paInt16
 class AudioHelper:
     def __init__(self):
         self.chunk_size = 1024
-        self.format = paInt16  # each sample is 16 bits, ie; 2 bytes
+        self.format = paInt16
         self.channels = 1
         self.rate = 44100
-
         self.pyaudio = None
         self.stream = None
         self.buf = bytearray()
 
-    def _start_stream(self):
+    def _start_stream(self, input=True, output=False):
         self.pyaudio = PyAudio()
         self.stream = self.pyaudio.open(
             format=self.format,
             channels=self.channels,
             rate=self.rate,
-            input=True,
+            input=input,
+            output=output,
             frames_per_buffer=self.chunk_size,
         )
 
     def _stop_stream(self):
+        if self.stream is None:
+            raise AttributeError(
+                "Stream is not started. Please call _start_stream first."
+            )
         self.stream.stop_stream()
         self.stream.close()
         self.pyaudio.terminate()
@@ -32,24 +36,17 @@ class AudioHelper:
     def _clear_buffer(self):
         self.buf.clear()
 
-    def record(self, time: int):
-        self._start_stream()
-        if self.stream is None:
-            raise AttributeError(
-                "Stream is not started. Please call _start_stream first."
-            )
+    def record(self, time: int) -> bytes:
+        self._start_stream(input=True, output=False)
         self._clear_buffer()
-        for _ in range(0, int(self.rate / self.chunk_size * time)):
+        iterations = self.rate // self.chunk_size * time
+        for _ in range(iterations):
             data = self.stream.read(self.chunk_size)
             self.buf.extend(data)
         self._stop_stream()
+        return bytes(self.buf)
 
-    def listen(self):
-        self._start_stream()
-        if self.stream is None:
-            raise AttributeError(
-                "Stream is not started. Please call _start_stream first."
-            )
-        data = self.stream.read(self.chunk_size)
+    def play(self, audio_data: bytes):
+        self._start_stream(input=False, output=True)
+        self.stream.write(audio_data)
         self._stop_stream()
-        return data
